# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Gateway — planned contract",
    "version": "0.1.0-design",
    "description": "Mixed implementation/design contract. Phase03 login and current-user endpoints are implemented; AI, conversation, and usage operations remain planned. Runtime FastAPI OpenAPI is authoritative for implemented routes. Metrics count logical requests separately from attempts and label unknown token metadata."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development server."
    }
  ],
  "security": [
    {
      "GatewayBearer": []
    }
  ],
...
  "paths": {
    "/v1/auth": {
      "post": {
        "operationId": "login",
        "tags": [
          "Auth"
        ],
        "summary": "Login a seeded user with JSON credentials.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            }
          },
          "401": {
            "description": "Invalid gateway credentials.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input or model refusal.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Local rate limit or daily quota.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Retry-After": {
                "description": "When applicable, seconds until a safe retry/window reset.",
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Redacted internal failure.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable, quota/configuration or persistence failure.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Retry-After": {
                "description": "When applicable, seconds until a safe retry/window reset.",
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "tags": [
          "Auth"
        ],
        "summary": "Return the identity associated with the bearer token.",
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            }
          },
          "401": {
            "description": "Invalid gateway credentials.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Redacted internal failure.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable, quota/configuration or persistence failure.",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated correlation UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Retry-After": {
                "description": "When applicable, seconds until a safe retry/window reset.",
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "GatewayBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AuthResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "type": "string",
            "minLength": 1
          },
          "token_type": {
            "type": "string",
            "enum": [
              "bearer"
            ]
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ]
      },
      "AuthenticatedUser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "request_id": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}